requirements.txt
flask
openai-whisper
groq
ffmpeg-python
torch
python-dotenv
